+++ cmd_util/program_environment.cpp
@@ -1,9 +1,10 @@
 #include <stdexcept>
 
-#include <program_environment.hpp>
+#include <util_helper/program_environment.hpp>
 #include <TDE/GCC_class.hpp>
 #include <TDE/GPS_class.hpp>
-#include <sigint.hpp>
+
+#include <util_helper/sigint.hpp>
 
 program_environment::program_environment() :
     _window_size(0),
